Moving backend folder to container
