+++ container/backend/backend.go
@@ -4,8 +4,9 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/opsfactory/kappa/backend/docker"
 	"github.com/opsfactory/kappa/config"
+	"github.com/opsfactory/kappa/container"
+	"github.com/opsfactory/kappa/container/backend/docker"
 	kappaevent "github.com/opsfactory/kappa/container/event"
 )
 
@@ -16,9 +17,8 @@ type Backend interface {
 
 func NewBackend(name string, c config.BackendConfig) (Backend, error) {
 	switch strings.ToLower(name) {
-	case "docker":
+	case container.DockerBackend:
 		return docker.NewDockerBackend(c)
 	}
-
 	return nil, fmt.Errorf("Unknown backend %s.", name)
 }
